Simplify CKAN installation
In CKAN 2.2 (and probably earlier) the package installation provides
default Apache configuration in /etc/apache2/sites-enabled. This was
what was actually being used by this module, which did not explicit
remove these files, while the configuration installed by
puppetlabs/apache was ignored (or at least listened on a different
port).

As such, there's no particular need to manage Apache and its
configuration with puppetlabs/apache. It introduces a lot of complexity
for no real gain.

So, this commit removes all of that stuff, and makes the job of the CKAN
module more-or-less installing CKAN and its dependencies.

A simple example manifest that uses this module might look like:

    node default {
      exec { "apt-get update":
        command => "/usr/bin/apt-get update",
      }

      Exec['apt-get update'] -> Package <| |>

      class { 'ckan':
        site_url              => 'test.ckan.com',
        site_title            => 'CKAN Test',
        site_description      => 'A shared environment for managing Data.',
        site_intro            => 'A CKAN test installation',
        site_about            => 'Pilot data catalogue and repository.',
        plugins               => 'stats text_preview recline_preview datastore resource_proxy pdf_preview',
        is_ckan_from_repo     => false,
        ckan_package_url      => 'http://packaging.ckan.org/python-ckan_2.2_amd64.deb',
        ckan_package_filename => 'python-ckan_2.2_amd64.deb',
      }
    }